Switch Algolia functions to thier own class
